refactor(model): model.py를 디렉토리로 이동 및 black 적용
